ml_microservice/requirements.txt
transformers~=4.36.2
torch~=2.1.2
confluent-kafka~=2.3.0
python-dotenv~=1.0.0